+++ examples/congestionchain/1_visualization.jl
@@ -45,13 +45,16 @@ scatter!(ax_fidZZ,ts,fidZZ,label="ZZ",color=(c2,0.1))
 display(fig)
 
 step_ts = range(0, 1000, step=0.1)
-for t in step_ts
+
+record(fig, "congestionchain.mp4", step_ts; framerate=50, visible=true) do t
     run(sim, t)
     ax.title = "t=$(t)"
-    notify(obs)
-    notify(ts)
-    autolimits!(ax_fidXX)
-    autolimits!(ax_fidZZ)
+    if length(ts[])>2 # to avoid failing autolimits on empty plots
+        notify(obs)
+        notify(ts)
+        autolimits!(ax_fidXX)
+        autolimits!(ax_fidZZ)
+    end
 end
 
 ##