+++ test/test_jet.jl
@@ -37,5 +37,6 @@ rep = report_package("QuantumSavory";
 
 @show length(JET.get_reports(rep))
 @show rep
-@test length(JET.get_reports(rep)) <= 133
+
+@test length(JET.get_reports(rep)) <= 146
 @test_broken length(JET.get_reports(rep)) == 0