+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # News
 
-## v0.5.0 - 2024-08-11
 
-- `observable` now takes a default value as a kwarg, i.e., you need to make the substitution `observable(regs, obs, 0.0; time)` ↦ `observable(regs, obs; something=0.0, time)`
+## v0.5.0 - 2024-09-05
+
+- Develop `CutoffProt` to deal with deadlocks in a simulation
+- Expand `SwapperProt` with `agelimit` to permit cutoff policies (with `CutoffProt`)
+- Tutorial and interactive examples for entanglement distribution on a grid with local-only knowledge
+- **(breaking)** `observable` now takes a default value as a kwarg, i.e., you need to make the substitution `observable(regs, obs, 0.0; time)` ↦ `observable(regs, obs; something=0.0, time)`
 - Bump QuantumSymbolics and QuantumOpticsBase compat bound and bump julia compat to 1.10.
+
+## v0.4.2 - 2024-08-13
+
+- Incorrect breaking release. It should have been 0.5 (see above).
 
 ## v0.4.1 - 2024-06-05